Add siwin examples and two windows (#22)
* siwin examples

* siwin examples

* siwin examples

* fix siwin examples

* fix siwin examples

* cleanup windyshim

* support metal in siwin

* support metal in siwin

* add siwin examples and cleanup

* add siwin examples

* add two windows

* add two windows

* add two windows

* add two windows

* fixup x11 and vulkan

* fixup x11 and vulkan

* cleanup siwin for vulkan

* fix software bit on vulkan

* include x11/wayland in siwin titles

* working on closing two windows

* updates

* updates

* fix tests

* fix tests and wayland/x11

* add FIGDRAW_FORCE_OPENGL

* add FIGDRAW_FORCE_OPENGL

* fix tests

* fix tests

* fix tests

* fix two windows

* fix two windows

* fix two windows

* update ci

* update ci

* update ci

* update ci

* update ci xwayland

+++ src/figdraw/opengl/glsl/atlas.vert
@@ -5,7 +5,7 @@ in vec2 vertexUv;
 in vec4 vertexColor;
 in vec4 vertexSdfParams;
 in vec4 vertexSdfRadii;
-in uint vertexSdfMode;
+in float vertexSdfMode;
 in vec2 vertexSdfFactors;
 
 uniform mat4 proj;
@@ -24,7 +24,7 @@ void main() {
   color = vertexColor;
   sdfParams = vertexSdfParams;
   sdfRadii = vertexSdfRadii;
-  sdfMode = float(vertexSdfMode);
+  sdfMode = vertexSdfMode;
   sdfFactors = vertexSdfFactors;
   gl_Position = proj * vec4(vertexPos.x, vertexPos.y, 0.0, 1.0);
 }
